Mades some fixes semanticListener

+++ src/edu/mit/compilers/ir/IrLiteralInt.java
@@ -3,9 +3,9 @@ package edu.mit.compilers.ir;
 import edu.mit.compilers.ScopeStack;
 
 public class IrLiteralInt extends IrLiteral {
-    private final int value;
+    private final long value;
 
-    public IrLiteralInt(int value, int lineNumber, int colNumber) {
+    public IrLiteralInt(long value, int lineNumber, int colNumber) {
         super(lineNumber, colNumber);
         this.value = value;
     }
